fix -Wshorten-64-to-32 warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38154 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/cart/retroreplay.c
@@ -1062,13 +1062,13 @@ static int retroreplay_common_attach(void)
     rr_enabled = 1;
 
     nofreeze_alarm = alarm_new(maincpu_alarm_context, "NoFreezeAlarm", nofreeze_alarm_handler, NULL);
-    
+
     return 0;
 }
 
 int retroreplay_bin_attach(const char *filename, uint8_t *rawcart)
 {
-    int len = 0;
+    size_t len = 0;
     FILE *fd;
 
     retroreplay_filetype = 0;
